fixed bugs from moodle report
fixed #9,  fixed #10

+++ delete_imglic_from_DB.php
@@ -28,5 +28,15 @@ $id = required_param('catid', PARAM_INT);
 
 if ($id && isloggedin() && has_capability('block/eexcess:myaddinstance', $systemcontext) && confirm_sesskey()) {
     $tablename = "block_eexcess_image_license";
-    $DB->delete_records($tablename, array("id" => $id));
+    $changedid = $DB->get_record($tablename, array("id" => $id), $fields = '*', $strictness = IGNORE_MISSING);
+    $useriddb = $changedid->userid;
+    $userid = $USER->id;
+    if ($useriddb === $userid) {
+        $DB->delete_records($tablename, array("id" => $id));
+        echo json_encode(array("success" => true));
+    } else {
+        echo json_encode(array("success" => false));
+    }
+} else {
+    echo json_encode(array("success" => false));
 }
